Moved factories to separate folders for cleaner structure.
