+++ backend/Factories/FacultyDataFactory.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Internships\Services;
+namespace Internships\Factories;
 
 use Internships\Models\Faculty;
 use Internships\Models\ValidationOptions;
